pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=77.0"]

[project]
name = "adk_cli"
version = "0.0.1"
license = "Apache-2.0"
license-files = ["LICENSE"]
description = "Project short description."
readme = "README.md"
authors = [{ name = "Allen Porter", email = "allen.porter@gmail.com" }]
requires-python = ">=3.14"
classifiers = []
dependencies = [
    "google-adk>=1.25.1",
    "click>=8.1.7",
    "rich>=13.7.0",
    "textual>=0.47.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.1",
    "pre-commit>=4.0.1",
    "coverage>=7.13.4",
    "pytest-asyncio>=1.3.0",
]

[project.scripts]
adk = "adk_cli.main:main"

[project.urls]
Source = "https://github.com/allenporter/adk_cli"

[tool.setuptools.packages.find]
include = ["adk_cli*"]

[tool.mypy]
exclude = [
  "venv/",
]
platform = "linux"
show_error_codes = true
follow_imports = "normal"
local_partial_types = true
strict_equality = true
no_implicit_optional = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
disable_error_code = [
    "import-untyped",
]
extra_checks = false
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
warn_return_any = true
warn_unreachable = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
